Move topics state to its own container
